Fix use integer to use 64-bit long, add interpreter integer op support
- Change integerShiftLeft/Right, integerDivide/Modulus, integerBitwiseNot
  from 32-bit int to 64-bit long (Perl 5 uses 64-bit IV)
- Add INTEGER_LEFT_SHIFT, INTEGER_RIGHT_SHIFT, INTEGER_DIV, INTEGER_MOD
  opcodes to interpreter backend with HINT_INTEGER detection
- Add INTEGER_*_ASSIGN compound assignment opcodes for interpreter
- Fix REPEAT opcode to handle RuntimeList count (constant sub returns)

+++ src/main/java/org/perlonjava/runtime/operators/MathOperators.java
@@ -434,16 +434,14 @@ public class MathOperators {
      */
     public static RuntimeScalar integerDivide(RuntimeScalar arg1, RuntimeScalar arg2) {
         // Convert to integers and perform integer division
-        int dividend = arg1.getInt();
-        int divisor = arg2.getInt();
+        long dividend = arg1.getLong();
+        long divisor = arg2.getLong();
 
-        // Handle division by zero
         if (divisor == 0) {
             throw new PerlCompilerException("Illegal division by zero");
         }
 
-        // Perform integer division
-        int result = dividend / divisor;
+        long result = dividend / divisor;
         return new RuntimeScalar(result);
     }
 
@@ -457,17 +455,14 @@ public class MathOperators {
      */
     public static RuntimeScalar integerModulus(RuntimeScalar arg1, RuntimeScalar arg2) {
         // Convert to integers and perform integer modulus
-        int dividend = arg1.getInt();
-        int divisor = arg2.getInt();
+        long dividend = arg1.getLong();
+        long divisor = arg2.getLong();
 
-        // Handle division by zero
         if (divisor == 0) {
             throw new PerlCompilerException("Illegal modulus zero");
         }
 
-        // In Perl with "use integer", modulus follows C99 truncated division rules
-        // The result has the sign of the dividend, not the divisor
-        int result = dividend % divisor;
+        long result = dividend % divisor;
         return new RuntimeScalar(result);
     }
